Fix api with DTO's and fix the endpoints with DTO's

+++ JobApplicationAssistantAPI/API/Controllers/ApplicationsController.cs
@@ -2,31 +2,39 @@
 using DAL.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using AutoMapper;
+using DAL.Dto;
 
 namespace API.Controllers
 {
     [ApiController]
-    [Route("api/[controller]")]
+    [Route("[controller]")]
     public class ApplicationsController : ControllerBase
     {
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IMapper _mapper;
 
-        public ApplicationsController(IUnitOfWork unitOfWork)
+        public ApplicationsController(IUnitOfWork unitOfWork, IMapper mapper)
         {
             _unitOfWork = unitOfWork;
+            _mapper = mapper;
         }
 
+        // GET: http://.../Applications
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Application>>> GetApplications()
+        public async Task<ActionResult<IEnumerable<ApplicationResponse>>> GetApplications()
         {
-            var applications = await _unitOfWork.ApplicationRepository.GetAllAsync(query =>
-                query.Include(a => a.Job)
-                     .ThenInclude(j => j.Company));
-            return Ok(applications);
+            var applications = await _unitOfWork.ApplicationRepository.GetAllAsync(
+                query => query.Include(a => a.Job)
+                              .ThenInclude(j => j.Company));
+
+            var applicationsDtos = _mapper.Map<IEnumerable<ApplicationResponse>>(applications);
+            return Ok(applicationsDtos);
         }
 
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Application>> GetApplication(int id)
+        // GET: http://.../Applications/by-id/{id}
+        [HttpGet("by-id/{id}")]
+        public async Task<ActionResult<Application>> GetApplicationById(int id)
         {
             var application = await _unitOfWork.ApplicationRepository.FindAsync(
                 a => a.Id == id,
@@ -34,33 +42,52 @@ namespace API.Controllers
                                 .ThenInclude(j => j.Company));
 
             if (application == null) return NotFound();
-            return Ok(application);
+
+            var applicationDto = _mapper.Map<ApplicationResponse>(application);
+            return Ok(applicationDto);
         }
-        [HttpPost]
-        public async Task<ActionResult<Application>> CreateApplication(Application application)
+
+        // POST: http://.../Applications/create
+        [HttpPost("create")]
+        public async Task<ActionResult<ApplicationRequest>> CreateApplication(ApplicationRequest applicationRequest)
         {
+            applicationRequest.Status = ApplicationStatus.Submitted;
+
+            var application = _mapper.Map<Application>(applicationRequest);
             application.ApplicationDate = DateTime.Now;
-            application.Status = ApplicationStatus.Submitted;
 
             await _unitOfWork.ApplicationRepository.InsertAsync(application);
             await _unitOfWork.SaveAsync();
-            return CreatedAtAction(nameof(GetApplication), new { id = application.Id }, application);
+
+            var appplicationResponse = _mapper.Map<ApplicationResponse>(application);
+            return CreatedAtAction(nameof(GetApplicationById), new { id = application.Id }, application);
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateApplication(int id, Application application)
+        // PUT: http://.../Applications/update/{id}
+        [HttpPut("update/{id}")]
+        public async Task<IActionResult> UpdateApplication(int id, [FromBody] ApplicationRequest applicationRequest)
         {
-            if (id != application.Id) return BadRequest();
-            await _unitOfWork.ApplicationRepository.UpdateAsync(application);
+            var existingApplication = await _unitOfWork.ApplicationRepository.GetByIDAsync(id);
+            if (existingApplication == null) return NotFound("Application not found.");
+
+            _mapper.Map(applicationRequest, existingApplication);
+
+            await _unitOfWork.ApplicationRepository.UpdateAsync(existingApplication);
             await _unitOfWork.SaveAsync();
+
             return NoContent();
         }
 
+        // DELETE: http://.../Applications/remove/{id}
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteApplication(int id)
+        public async Task<IActionResult> RemoveApplication(int id)
         {
+            var existingApplication = await _unitOfWork.ApplicationRepository.GetByIDAsync(id);
+            if (existingApplication == null) return NotFound("Application not found.");
+
             await _unitOfWork.ApplicationRepository.DeleteAsync(id);
             await _unitOfWork.SaveAsync();
+
             return NoContent();
         }
